Change from cdcacm to vendor specific
Switch from 8 interfaces to 4
Use NovaTech VID and 124 for product ID
