--- src/cdcacm/Makefile
@@ -1,8 +1,0 @@
-CFLAGS += \
-	-Icdcacm
-
-SRC += \
-	cdcacm/cdcacm.c
-
-VPATH += \
-	cdcacm
