--- src/cdcacm/cdcacm.c
@@ -1,648 +1,0 @@
-/*
- * This file is part of the Black Magic Debug project.
- *
- * Copyright (C) 2011  Black Sphere Technologies Ltd.
- * Written by Gareth McMullin <gareth@blacksphere.co.nz>
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-/* This file implements a the USB Communications Device Class - Abstract
- * Control Model (CDC-ACM) as defined in CDC PSTN subclass 1.2.
- * A Device Firmware Upgrade (DFU 1.1) class interface is provided for
- * field firmware upgrade.
- *
- * The device's unique id is used as the USB serial number string.
- */
-
-#include <stdlib.h>
-
-#include <libopencm3/cm3/nvic.h>
-#include <libopencm3/usb/usbd.h>
-#include <libopencm3/usb/cdc.h>
-#include <libopencm3/cm3/scb.h>
-#include <libopencm3/stm32/rcc.h>
-
-#include <uart.h>
-
-#include "platform.h"
-#include "cdcacm.h"
-
-#define BOARD_IDENT             "NovaTech USB to serial"
-
-#define SERIALNO_FLASH_LOCATION	0x8001ff0
-
-extern struct uart_t uarts[];
-
-usbd_device * usbdev;
-
-static char *get_dev_unique_id(char *serial_no);
-
-static int configured;
-
-static const struct usb_device_descriptor dev = {
-        .bLength = USB_DT_DEVICE_SIZE,
-        .bDescriptorType = USB_DT_DEVICE,
-        .bcdUSB = 0x0200,
-        .bDeviceClass = 0xEF,		/* Miscellaneous Device */
-        .bDeviceSubClass = 2,		/* Common Class */
-        .bDeviceProtocol = 1,		/* Interface Association */
-        .bMaxPacketSize0 = 64,
-        .idVendor = 0x1D50,
-        .idProduct = 0x6018,
-        .bcdDevice = 0x0100,
-        .iManufacturer = 1,
-        .iProduct = 2,
-        .iSerialNumber = 3,
-        .bNumConfigurations = 1,
-};
-
-static const struct usb_endpoint_descriptor port1_data_endp[] = {{
-	.bLength = USB_DT_ENDPOINT_SIZE,
-	.bDescriptorType = USB_DT_ENDPOINT,
-	.bEndpointAddress = 0x01,
-	.bmAttributes = USB_ENDPOINT_ATTR_BULK,
-	.wMaxPacketSize = CDCACM_PACKET_SIZE,
-	.bInterval = 1,
-}, {
-	.bLength = USB_DT_ENDPOINT_SIZE,
-	.bDescriptorType = USB_DT_ENDPOINT,
-	.bEndpointAddress = 0x81,
-	.bmAttributes = USB_ENDPOINT_ATTR_BULK,
-	.wMaxPacketSize = CDCACM_PACKET_SIZE,
-	.bInterval = 1,
-}};
-
-static const struct {
-	struct usb_cdc_header_descriptor header;
-	struct usb_cdc_call_management_descriptor call_mgmt;
-	struct usb_cdc_acm_descriptor acm;
-	struct usb_cdc_union_descriptor cdc_union;
-} __attribute__((packed)) port1_cdcacm_functional_descriptors = {
-	.header = {
-		.bFunctionLength = sizeof(struct usb_cdc_header_descriptor),
-		.bDescriptorType = CS_INTERFACE,
-		.bDescriptorSubtype = USB_CDC_TYPE_HEADER,
-		.bcdCDC = 0x0110,
-	},
-	.call_mgmt = {
-		.bFunctionLength =
-			sizeof(struct usb_cdc_call_management_descriptor),
-		.bDescriptorType = CS_INTERFACE,
-		.bDescriptorSubtype = USB_CDC_TYPE_CALL_MANAGEMENT,
-		.bmCapabilities = 0,
-		.bDataInterface = 1,
-	},
-	.acm = {
-		.bFunctionLength = sizeof(struct usb_cdc_acm_descriptor),
-		.bDescriptorType = CS_INTERFACE,
-		.bDescriptorSubtype = USB_CDC_TYPE_ACM,
-		.bmCapabilities = 2, /* SET_LINE_CODING supported */
-	},
-	.cdc_union = {
-		.bFunctionLength = sizeof(struct usb_cdc_union_descriptor),
-		.bDescriptorType = CS_INTERFACE,
-		.bDescriptorSubtype = USB_CDC_TYPE_UNION,
-		.bControlInterface = 0,
-		.bSubordinateInterface0 = 1,
-	 }
-};
-
-static const struct usb_interface_descriptor port1_comm_iface[] = {{
-	.bLength = USB_DT_INTERFACE_SIZE,
-	.bDescriptorType = USB_DT_INTERFACE,
-	.bInterfaceNumber = 0,
-	.bAlternateSetting = 0,
-	.bNumEndpoints = 0,
-	.bInterfaceClass = USB_CLASS_CDC,
-	.bInterfaceSubClass = USB_CDC_SUBCLASS_ACM,
-	.bInterfaceProtocol = USB_CDC_PROTOCOL_AT,
-	.iInterface = 4,
-
-	.extra = &port1_cdcacm_functional_descriptors,
-	.extralen = sizeof(port1_cdcacm_functional_descriptors)
-}};
-
-static const struct usb_interface_descriptor port1_data_iface[] = {{
-	.bLength = USB_DT_INTERFACE_SIZE,
-	.bDescriptorType = USB_DT_INTERFACE,
-	.bInterfaceNumber = 1,
-	.bAlternateSetting = 0,
-	.bNumEndpoints = 2,
-	.bInterfaceClass = USB_CLASS_DATA,
-	.bInterfaceSubClass = 0,
-	.bInterfaceProtocol = 0,
-	.iInterface = 0,
-
-	.endpoint = port1_data_endp,
-}};
-
-static const struct usb_iface_assoc_descriptor port1_assoc = {
-	.bLength = USB_DT_INTERFACE_ASSOCIATION_SIZE,
-	.bDescriptorType = USB_DT_INTERFACE_ASSOCIATION,
-	.bFirstInterface = 0,
-	.bInterfaceCount = 2,
-	.bFunctionClass = USB_CLASS_CDC,
-	.bFunctionSubClass = USB_CDC_SUBCLASS_ACM,
-	.bFunctionProtocol = USB_CDC_PROTOCOL_AT,
-	.iFunction = 0,
-};
-
-/* Serial ACM interface */
-static const struct usb_endpoint_descriptor port2_data_endp[] = {{
-	.bLength = USB_DT_ENDPOINT_SIZE,
-	.bDescriptorType = USB_DT_ENDPOINT,
-	.bEndpointAddress = 0x03,
-	.bmAttributes = USB_ENDPOINT_ATTR_BULK,
-	.wMaxPacketSize = CDCACM_PACKET_SIZE,
-	.bInterval = 1,
-}, {
-	.bLength = USB_DT_ENDPOINT_SIZE,
-	.bDescriptorType = USB_DT_ENDPOINT,
-	.bEndpointAddress = 0x83,
-	.bmAttributes = USB_ENDPOINT_ATTR_BULK,
-	.wMaxPacketSize = CDCACM_PACKET_SIZE,
-	.bInterval = 1,
-}};
-
-static const struct {
-	struct usb_cdc_header_descriptor header;
-	struct usb_cdc_call_management_descriptor call_mgmt;
-	struct usb_cdc_acm_descriptor acm;
-	struct usb_cdc_union_descriptor cdc_union;
-} __attribute__((packed)) port2_cdcacm_functional_descriptors = {
-	.header = {
-		.bFunctionLength = sizeof(struct usb_cdc_header_descriptor),
-		.bDescriptorType = CS_INTERFACE,
-		.bDescriptorSubtype = USB_CDC_TYPE_HEADER,
-		.bcdCDC = 0x0110,
-	},
-	.call_mgmt = {
-		.bFunctionLength =
-			sizeof(struct usb_cdc_call_management_descriptor),
-		.bDescriptorType = CS_INTERFACE,
-		.bDescriptorSubtype = USB_CDC_TYPE_CALL_MANAGEMENT,
-		.bmCapabilities = 0,
-		.bDataInterface = 3,
-	},
-	.acm = {
-		.bFunctionLength = sizeof(struct usb_cdc_acm_descriptor),
-		.bDescriptorType = CS_INTERFACE,
-		.bDescriptorSubtype = USB_CDC_TYPE_ACM,
-		.bmCapabilities = 2, /* SET_LINE_CODING supported*/
-	},
-	.cdc_union = {
-		.bFunctionLength = sizeof(struct usb_cdc_union_descriptor),
-		.bDescriptorType = CS_INTERFACE,
-		.bDescriptorSubtype = USB_CDC_TYPE_UNION,
-		.bControlInterface = 2,
-		.bSubordinateInterface0 = 3,
-	 }
-};
-
-static const struct usb_interface_descriptor port2_comm_iface[] = {{
-	.bLength = USB_DT_INTERFACE_SIZE,
-	.bDescriptorType = USB_DT_INTERFACE,
-	.bInterfaceNumber = 2,
-	.bAlternateSetting = 0,
-	.bNumEndpoints = 0,
-	.bInterfaceClass = USB_CLASS_CDC,
-	.bInterfaceSubClass = USB_CDC_SUBCLASS_ACM,
-	.bInterfaceProtocol = USB_CDC_PROTOCOL_AT,
-	.iInterface = 5,
-
-	.extra = &port2_cdcacm_functional_descriptors,
-	.extralen = sizeof(port2_cdcacm_functional_descriptors)
-}};
-
-static const struct usb_interface_descriptor port2_data_iface[] = {{
-	.bLength = USB_DT_INTERFACE_SIZE,
-	.bDescriptorType = USB_DT_INTERFACE,
-	.bInterfaceNumber = 3,
-	.bAlternateSetting = 0,
-	.bNumEndpoints = 2,
-	.bInterfaceClass = USB_CLASS_DATA,
-	.bInterfaceSubClass = 0,
-	.bInterfaceProtocol = 0,
-	.iInterface = 0,
-
-	.endpoint = port2_data_endp,
-}};
-
-static const struct usb_iface_assoc_descriptor port2_assoc = {
-	.bLength = USB_DT_INTERFACE_ASSOCIATION_SIZE,
-	.bDescriptorType = USB_DT_INTERFACE_ASSOCIATION,
-	.bFirstInterface = 2,
-	.bInterfaceCount = 2,
-	.bFunctionClass = USB_CLASS_CDC,
-	.bFunctionSubClass = USB_CDC_SUBCLASS_ACM,
-	.bFunctionProtocol = USB_CDC_PROTOCOL_AT,
-	.iFunction = 0,
-};
-
-/* Serial ACM interface */
-static const struct usb_endpoint_descriptor port3_data_endp[] = {{
-	.bLength = USB_DT_ENDPOINT_SIZE,
-	.bDescriptorType = USB_DT_ENDPOINT,
-	.bEndpointAddress = 0x05,
-	.bmAttributes = USB_ENDPOINT_ATTR_BULK,
-	.wMaxPacketSize = CDCACM_PACKET_SIZE,
-	.bInterval = 1,
-}, {
-	.bLength = USB_DT_ENDPOINT_SIZE,
-	.bDescriptorType = USB_DT_ENDPOINT,
-	.bEndpointAddress = 0x85,
-	.bmAttributes = USB_ENDPOINT_ATTR_BULK,
-	.wMaxPacketSize = CDCACM_PACKET_SIZE,
-	.bInterval = 1,
-}};
-
-static const struct {
-	struct usb_cdc_header_descriptor header;
-	struct usb_cdc_call_management_descriptor call_mgmt;
-	struct usb_cdc_acm_descriptor acm;
-	struct usb_cdc_union_descriptor cdc_union;
-} __attribute__((packed)) port3_cdcacm_functional_descriptors = {
-	.header = {
-		.bFunctionLength = sizeof(struct usb_cdc_header_descriptor),
-		.bDescriptorType = CS_INTERFACE,
-		.bDescriptorSubtype = USB_CDC_TYPE_HEADER,
-		.bcdCDC = 0x0110,
-	},
-	.call_mgmt = {
-		.bFunctionLength =
-			sizeof(struct usb_cdc_call_management_descriptor),
-		.bDescriptorType = CS_INTERFACE,
-		.bDescriptorSubtype = USB_CDC_TYPE_CALL_MANAGEMENT,
-		.bmCapabilities = 0,
-		.bDataInterface = 5,
-	},
-	.acm = {
-		.bFunctionLength = sizeof(struct usb_cdc_acm_descriptor),
-		.bDescriptorType = CS_INTERFACE,
-		.bDescriptorSubtype = USB_CDC_TYPE_ACM,
-		.bmCapabilities = 2, /* SET_LINE_CODING supported*/
-	},
-	.cdc_union = {
-		.bFunctionLength = sizeof(struct usb_cdc_union_descriptor),
-		.bDescriptorType = CS_INTERFACE,
-		.bDescriptorSubtype = USB_CDC_TYPE_UNION,
-		.bControlInterface = 4,
-		.bSubordinateInterface0 = 5,
-	 }
-};
-
-static const struct usb_interface_descriptor port3_comm_iface[] = {{
-	.bLength = USB_DT_INTERFACE_SIZE,
-	.bDescriptorType = USB_DT_INTERFACE,
-	.bInterfaceNumber = 4,
-	.bAlternateSetting = 0,
-	.bNumEndpoints = 0,
-	.bInterfaceClass = USB_CLASS_CDC,
-	.bInterfaceSubClass = USB_CDC_SUBCLASS_ACM,
-	.bInterfaceProtocol = USB_CDC_PROTOCOL_AT,
-	.iInterface = 6,
-
-	.extra = &port3_cdcacm_functional_descriptors,
-	.extralen = sizeof(port3_cdcacm_functional_descriptors)
-}};
-
-static const struct usb_interface_descriptor port3_data_iface[] = {{
-	.bLength = USB_DT_INTERFACE_SIZE,
-	.bDescriptorType = USB_DT_INTERFACE,
-	.bInterfaceNumber = 5,
-	.bAlternateSetting = 0,
-	.bNumEndpoints = 2,
-	.bInterfaceClass = USB_CLASS_DATA,
-	.bInterfaceSubClass = 0,
-	.bInterfaceProtocol = 0,
-	.iInterface = 0,
-
-	.endpoint = port3_data_endp,
-}};
-
-static const struct usb_iface_assoc_descriptor port3_assoc = {
-	.bLength = USB_DT_INTERFACE_ASSOCIATION_SIZE,
-	.bDescriptorType = USB_DT_INTERFACE_ASSOCIATION,
-	.bFirstInterface = 4,
-	.bInterfaceCount = 2,
-	.bFunctionClass = USB_CLASS_CDC,
-	.bFunctionSubClass = USB_CDC_SUBCLASS_ACM,
-	.bFunctionProtocol = USB_CDC_PROTOCOL_AT,
-	.iFunction = 0,
-};
-
-/* Serial ACM interface */
-static const struct usb_endpoint_descriptor port4_data_endp[] = {{
-	.bLength = USB_DT_ENDPOINT_SIZE,
-	.bDescriptorType = USB_DT_ENDPOINT,
-	.bEndpointAddress = 0x07,
-	.bmAttributes = USB_ENDPOINT_ATTR_BULK,
-	.wMaxPacketSize = CDCACM_PACKET_SIZE,
-	.bInterval = 1,
-}, {
-	.bLength = USB_DT_ENDPOINT_SIZE,
-	.bDescriptorType = USB_DT_ENDPOINT,
-	.bEndpointAddress = 0x87,
-	.bmAttributes = USB_ENDPOINT_ATTR_BULK,
-	.wMaxPacketSize = CDCACM_PACKET_SIZE,
-	.bInterval = 1,
-}};
-
-static const struct {
-	struct usb_cdc_header_descriptor header;
-	struct usb_cdc_call_management_descriptor call_mgmt;
-	struct usb_cdc_acm_descriptor acm;
-	struct usb_cdc_union_descriptor cdc_union;
-} __attribute__((packed)) port4_cdcacm_functional_descriptors = {
-	.header = {
-		.bFunctionLength = sizeof(struct usb_cdc_header_descriptor),
-		.bDescriptorType = CS_INTERFACE,
-		.bDescriptorSubtype = USB_CDC_TYPE_HEADER,
-		.bcdCDC = 0x0110,
-	},
-	.call_mgmt = {
-		.bFunctionLength =
-			sizeof(struct usb_cdc_call_management_descriptor),
-		.bDescriptorType = CS_INTERFACE,
-		.bDescriptorSubtype = USB_CDC_TYPE_CALL_MANAGEMENT,
-		.bmCapabilities = 0,
-		.bDataInterface = 7,
-	},
-	.acm = {
-		.bFunctionLength = sizeof(struct usb_cdc_acm_descriptor),
-		.bDescriptorType = CS_INTERFACE,
-		.bDescriptorSubtype = USB_CDC_TYPE_ACM,
-		.bmCapabilities = 2, /* SET_LINE_CODING supported*/
-	},
-	.cdc_union = {
-		.bFunctionLength = sizeof(struct usb_cdc_union_descriptor),
-		.bDescriptorType = CS_INTERFACE,
-		.bDescriptorSubtype = USB_CDC_TYPE_UNION,
-		.bControlInterface = 6,
-		.bSubordinateInterface0 = 7,
-	 }
-};
-
-static const struct usb_interface_descriptor port4_comm_iface[] = {{
-	.bLength = USB_DT_INTERFACE_SIZE,
-	.bDescriptorType = USB_DT_INTERFACE,
-	.bInterfaceNumber = 6,
-	.bAlternateSetting = 0,
-	.bNumEndpoints = 0,
-	.bInterfaceClass = USB_CLASS_CDC,
-	.bInterfaceSubClass = USB_CDC_SUBCLASS_ACM,
-	.bInterfaceProtocol = USB_CDC_PROTOCOL_AT,
-	.iInterface = 7,
-
-	.extra = &port4_cdcacm_functional_descriptors,
-	.extralen = sizeof(port4_cdcacm_functional_descriptors)
-}};
-
-static const struct usb_interface_descriptor port4_data_iface[] = {{
-	.bLength = USB_DT_INTERFACE_SIZE,
-	.bDescriptorType = USB_DT_INTERFACE,
-	.bInterfaceNumber = 7,
-	.bAlternateSetting = 0,
-	.bNumEndpoints = 2,
-	.bInterfaceClass = USB_CLASS_DATA,
-	.bInterfaceSubClass = 0,
-	.bInterfaceProtocol = 0,
-	.iInterface = 0,
-
-	.endpoint = port4_data_endp,
-}};
-
-static const struct usb_iface_assoc_descriptor port4_assoc = {
-	.bLength = USB_DT_INTERFACE_ASSOCIATION_SIZE,
-	.bDescriptorType = USB_DT_INTERFACE_ASSOCIATION,
-	.bFirstInterface = 6,
-	.bInterfaceCount = 2,
-	.bFunctionClass = USB_CLASS_CDC,
-	.bFunctionSubClass = USB_CDC_SUBCLASS_ACM,
-	.bFunctionProtocol = USB_CDC_PROTOCOL_AT,
-	.iFunction = 0,
-};
-
-static const struct usb_interface ifaces[] = {{
-	.num_altsetting = 1,
-	.iface_assoc = &port1_assoc,
-	.altsetting = port1_comm_iface,
-}, {
-	.num_altsetting = 1,
-	.altsetting = port1_data_iface,
-}, {
-	.num_altsetting = 1,
-	.iface_assoc = &port2_assoc,
-	.altsetting = port2_comm_iface,
-}, {
-	.num_altsetting = 1,
-	.altsetting = port2_data_iface,
-}, {
-	.num_altsetting = 1,
-	.iface_assoc = &port3_assoc,
-	.altsetting = port3_comm_iface,
-}, {
-	.num_altsetting = 1,
-	.altsetting = port3_data_iface,
-}, {
-	.num_altsetting = 1,
-	.iface_assoc = &port4_assoc,
-	.altsetting = port4_comm_iface,
-}, {
-	.num_altsetting = 1,
-	.altsetting = port4_data_iface,
-}};
-
-static const struct usb_config_descriptor config = {
-	.bLength = USB_DT_CONFIGURATION_SIZE,
-	.bDescriptorType = USB_DT_CONFIGURATION,
-	.wTotalLength = 0,
-	.bNumInterfaces = 8,
-	.bConfigurationValue = 1,
-	.iConfiguration = 0,
-	.bmAttributes = 0x80,
-	.bMaxPower = 0x32,
-
-	.interface = ifaces,
-};
-
-char serial_no[9];
-
-static const char *usb_strings[] = {
-	"NovaTech LLC",
-	BOARD_IDENT,
-	(char *)SERIALNO_FLASH_LOCATION,
-	"Port 1",
-	"Port 2",
-	"Port 3",
-	"Port 4",
-};
-
-static int cdcacm_control_request(usbd_device *dev,
-		struct usb_setup_data *req, uint8_t **buf, uint16_t *len,
-		void (**complete)(usbd_device *dev, struct usb_setup_data *req))
-{
-	(void)dev;
-	(void)complete;
-	(void)buf;
-	(void)len;
-
-	switch(req->bRequest) {
-	case USB_CDC_REQ_SET_CONTROL_LINE_STATE:
-		switch(req->wIndex) {
-		case 6:
-			usbuart_set_control_line_state(&uarts[ACM3_UART_INDEX], req->wValue);
-			return 1;
-		case 4:
-			usbuart_set_control_line_state(&uarts[ACM2_UART_INDEX], req->wValue);
-			return 1;
-		case 2:
-			usbuart_set_control_line_state(&uarts[ACM1_UART_INDEX], req->wValue);
-			return 1;
-		case 0:
-			usbuart_set_control_line_state(&uarts[ACM0_UART_INDEX], req->wValue);
-			return 1;
-		default:
-			return 0;
-		}
-	case USB_CDC_REQ_SET_LINE_CODING:
-		if(*len < sizeof(struct usb_cdc_line_coding))
-			return 0;
-
-		switch(req->wIndex) {
-		case 6:
-			usbuart_set_line_coding(&uarts[ACM3_UART_INDEX], (struct usb_cdc_line_coding*)*buf);
-			return 1;
-		case 4:
-			usbuart_set_line_coding(&uarts[ACM2_UART_INDEX], (struct usb_cdc_line_coding*)*buf);
-			return 1;
-		case 2:
-			usbuart_set_line_coding(&uarts[ACM1_UART_INDEX], (struct usb_cdc_line_coding*)*buf);
-			return 1;
-		case 0:
-			usbuart_set_line_coding(&uarts[ACM0_UART_INDEX], (struct usb_cdc_line_coding*)*buf);
-			return 1;
-		default:
-			return 0;
-		}
-	case USB_CDC_SET_FLOW_CONTROL:
-		switch(req->wIndex) {
-		case 6:
-			usbuart_set_flow_control(&uarts[ACM3_UART_INDEX], req->wValue);
-			return 1;
-		case 4:
-			usbuart_set_flow_control(&uarts[ACM2_UART_INDEX], req->wValue);
-			return 1;
-		case 2:
-			usbuart_set_flow_control(&uarts[ACM1_UART_INDEX], req->wValue);
-			return 1;
-		case 0:
-			usbuart_set_flow_control(&uarts[ACM0_UART_INDEX], req->wValue);
-			return 1;
-		default:
-			return 0;
-		}
-	}
-	return 0;
-}
-
-int cdcacm_get_config(void)
-{
-	return configured;
-}
-
-static void cdcacm_set_config(usbd_device *dev, uint16_t wValue)
-{
-	configured = wValue;
-
-	/* Serial interface */
-	usbd_ep_setup(dev, 0x01, USB_ENDPOINT_ATTR_BULK,
-					CDCACM_PACKET_SIZE, ACM0_UART_OUT_CALL);
-	usbd_ep_setup(dev, 0x81, USB_ENDPOINT_ATTR_BULK,
-					CDCACM_PACKET_SIZE, ACM0_UART_IN_CALL);
-
-	/* Serial interface */
-	usbd_ep_setup(dev, 0x03, USB_ENDPOINT_ATTR_BULK,
-					CDCACM_PACKET_SIZE, ACM1_UART_OUT_CALL);
-	usbd_ep_setup(dev, 0x83, USB_ENDPOINT_ATTR_BULK,
-					CDCACM_PACKET_SIZE, ACM1_UART_IN_CALL);
-	
-	/* Serial interface */
-	usbd_ep_setup(dev, 0x05, USB_ENDPOINT_ATTR_BULK,
-					CDCACM_PACKET_SIZE, ACM2_UART_OUT_CALL);
-	usbd_ep_setup(dev, 0x85, USB_ENDPOINT_ATTR_BULK,
-					CDCACM_PACKET_SIZE, ACM2_UART_IN_CALL);
-	
-	/* Serial interface */
-	usbd_ep_setup(dev, 0x07, USB_ENDPOINT_ATTR_BULK,
-					CDCACM_PACKET_SIZE, ACM3_UART_OUT_CALL);
-	usbd_ep_setup(dev, 0x87, USB_ENDPOINT_ATTR_BULK,
-					CDCACM_PACKET_SIZE, ACM3_UART_IN_CALL);
-
-
-	usbd_register_control_callback(dev,
-			USB_REQ_TYPE_CLASS | USB_REQ_TYPE_INTERFACE,
-			USB_REQ_TYPE_TYPE | USB_REQ_TYPE_RECIPIENT,
-			cdcacm_control_request);
-}
-
-/* We need a special large control buffer for this device: */
-uint8_t usbd_control_buffer[256];
-
-static char *get_dev_unique_id(char *s)
-{
-	uint32_t unique_id = *unique_id_p +
-			*(unique_id_p + 1) +
-			*(unique_id_p + 2);
-        int i;
-
-        /* Fetch serial number from chip's unique ID */
-        for(i = 0; i < 8; i++) {
-                s[7-i] = ((unique_id >> (4*i)) & 0xF) + '0';
-        }
-        for(i = 0; i < 8; i++)
-                if(s[i] > '9')
-                        s[i] += 'A' - '9' - 1;
-	s[8] = 0;
-
-	return s;
-}
-
-void cdcacm_init(void) {
-	// reset that the USB is not configured
-	configured = 0;
-	// reset USB hardware and re-enumerate USB
-	usb_platform_init();
-
-	// Initilize the USB stack
-	get_dev_unique_id(serial_no);
-	usbdev = usbd_init(&USB_DRIVER, &dev, &config, usb_strings,
-			    sizeof(usb_strings)/sizeof(char *),
-			    usbd_control_buffer, sizeof(usbd_control_buffer));
-	usbd_register_set_config_callback(usbdev, cdcacm_set_config);
-	
-	// setup the ISR
-	nvic_set_priority(USB_IRQ, IRQ_PRI_USB);
-	nvic_enable_irq(USB_IRQ);
-}
-
-
-/*
- * Interrupt routines
- */
-
-void USB_ISR(void) {
-	usbd_poll(usbdev);
-}
