+++ src/usb/Makefile
@@ -1,0 +1,8 @@
+CFLAGS += \
+	-Iusb
+
+SRC += \
+	usb/usb.c
+
+VPATH += \
+	usb
